lustre: obdclass: restore md_stats code
Originally the linux lustre client contained the md_stats code
but no one used it at that time so some parts of the code was
removed. The remaining md_stats code is layered in many levels
of macros that make the code difficult to read as well as
introduce undetected errors. This patch peels away the macro
wrappers by replacing it with the function lprocfs_count_incr()
which doesn't care about the order of the function pointers in
struct md_ops as the original code did.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-bug-id: https://jira.whamcloud.com/browse/LU-9855
Reviewed-on: https://review.whamcloud.com/32822
Reviewed-by: Ben Evans <bevans@cray.com>
Reviewed-by: John L. Hammond <jhammond@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>